+++ net/sched/cls_api.c
@@ -1252,7 +1252,7 @@ errout_qdisc:
 	return err;
 }
 
-static int __tcf_qdisc_cl_find(struct Qdisc *q, u32 parent, unsigned long *cl,
+static int __tcf_qdisc_cl_find(struct Qdisc *q, u32 parent, uintptr_t *cl,
 			       int ifindex, struct netlink_ext_ack *extack)
 {
 	if (ifindex == TCM_IFINDEX_MAGIC_BLOCK)
@@ -1273,7 +1273,7 @@ static int __tcf_qdisc_cl_find(struct Qdisc *q, u32 parent, unsigned long *cl,
 }
 
 static struct tcf_block *__tcf_block_find(struct net *net, struct Qdisc *q,
-					  unsigned long cl, int ifindex,
+					  uintptr_t cl, int ifindex,
 					  u32 block_index,
 					  struct netlink_ext_ack *extack)
 {
@@ -1347,7 +1347,7 @@ static void tcf_block_refcnt_put(struct tcf_block *block, bool rtnl_held)
  */
 
 static struct tcf_block *tcf_block_find(struct net *net, struct Qdisc **q,
-					u32 *parent, unsigned long *cl,
+					u32 *parent, uintptr_t *cl,
 					int ifindex, u32 block_index,
 					struct netlink_ext_ack *extack)
 {
@@ -2250,7 +2250,7 @@ static int tc_new_tfilter(struct sk_buff *skb, struct nlmsghdr *n,
 	struct tcf_chain *chain;
 	struct tcf_block *block;
 	struct tcf_proto *tp;
-	unsigned long cl;
+	uintptr_t cl;
 	void *fh;
 	int err;
 	int tp_created;
@@ -2485,7 +2485,7 @@ static int tc_del_tfilter(struct sk_buff *skb, struct nlmsghdr *n,
 	struct tcf_chain *chain = NULL;
 	struct tcf_block *block = NULL;
 	struct tcf_proto *tp = NULL;
-	unsigned long cl = 0;
+	uintptr_t cl = 0;
 	void *fh = NULL;
 	int err;
 	bool rtnl_held = false;
@@ -2644,7 +2644,7 @@ static int tc_get_tfilter(struct sk_buff *skb, struct nlmsghdr *n,
 	struct tcf_chain *chain = NULL;
 	struct tcf_block *block = NULL;
 	struct tcf_proto *tp = NULL;
-	unsigned long cl = 0;
+	uintptr_t cl = 0;
 	void *fh = NULL;
 	int err;
 	bool rtnl_held = false;
@@ -2884,7 +2884,7 @@ static int tc_dump_tfilter(struct sk_buff *skb, struct netlink_callback *cb)
 	} else {
 		const struct Qdisc_class_ops *cops;
 		struct net_device *dev;
-		unsigned long cl = 0;
+		uintptr_t cl = 0;
 
 		dev = __dev_get_by_index(net, tcm->tcm_ifindex);
 		if (!dev)
@@ -3121,7 +3121,7 @@ static int tc_ctl_chain(struct sk_buff *skb, struct nlmsghdr *n,
 	struct Qdisc *q;
 	struct tcf_chain *chain;
 	struct tcf_block *block;
-	unsigned long cl;
+	uintptr_t cl;
 	int err;
 
 replay:
@@ -3269,7 +3269,7 @@ static int tc_dump_chain(struct sk_buff *skb, struct netlink_callback *cb)
 	} else {
 		const struct Qdisc_class_ops *cops;
 		struct net_device *dev;
-		unsigned long cl = 0;
+		uintptr_t cl = 0;
 
 		dev = __dev_get_by_index(net, tcm->tcm_ifindex);
 		if (!dev)
